fix: fixing the routing  of initial page and the login page
