--- src/app/entrar/page.js
@@ -1,8 +1,0 @@
-import Login from "@/components/Login";
-
-
-export default function Entrar(){
-    return(
-        <Login />
-    );
-}
